fix: #1891 Re-arrange repeat/repeatForever api + new tests
