+++ src/engine/Actions/Index.ts
@@ -1,6 +1,10 @@
 export * from './ActionContext';
+export * from './ActionQueue';
 export * from './Actionable';
 export * from './RotationType';
+
+export * from './Repeat';
+export * from './RepeatForever';
 
 import * as actions from './Action';
 export { actions as Actions };
